backend/requirements.txt
# FastAPI Backend Dependencies for AI Bug Analyzer
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
sqlalchemy>=2.0.0
sqlite3  # Built-in Python module
python-dotenv>=1.0.0
requests>=2.31.0
python-multipart>=0.0.6
httpx>=0.25.0
aiohttp>=3.9.0
asyncio-mqtt>=0.13.0
python-dateutil>=2.8.2

# AI and text processing
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.1.0

# CORS and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
